+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - fix: cancelling a workshop collection creation will crash the program
 - fix: avoid circular references in linked workshop collections
 - add: moving addons display
+- add: support Ctrl+X and Ctrl+V to move addons and support mouse side button to goto the parent group
 # v0.5.0
 - add: workshop collection creation
 - add: auto set the name of the workshop addon after download